requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.1.0
openai>=1.3.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
